Merge the latest code into master

+++ exchangis-job/exchangis-job-server/src/main/java/com/webank/wedatasphere/exchangis/job/server/service/impl/ExchangisJobDsBindServiceImpl.java
@@ -26,14 +26,13 @@ public class ExchangisJobDsBindServiceImpl implements ExchangisJobDsBindService 
         for (ExchangisJobDsBind dsBind : dsBinds) {
             this.dsBindMapper.insert(dsBind);
         }
-
     }
 
     @Override
     public boolean inUse(Long datasourceId) {
         QueryWrapper<ExchangisJobDsBind> condition = new QueryWrapper<>();
         condition.eq("source_ds_id", datasourceId).or().eq("sink_ds_id", datasourceId);
-        Integer count = Optional.ofNullable(this.dsBindMapper.selectCount(condition)).orElse(0);
+        Long count = Optional.ofNullable(this.dsBindMapper.selectCount(condition)).orElse(0L);
         return count > 0;
     }
 }
